SEANewsAlert/pyproject.toml
[project]
name = "seasnews-ai-system"
version = "2.0.0"
description = "東南亞金融海外新聞智能搜尋與報告系統 - FastAPI 版本"
requires-python = ">=3.11"
dependencies = [
    "agno>=2.0.0",
    "openai>=1.0.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.6.0",
    "pydantic[email]>=2.6.0",
    "streamlit>=1.32.0",
    "python-dotenv>=1.0.0",
    "reportlab>=4.0.0",
    "markdown>=3.5.0",
    "weasyprint>=60.0",
    "aiosmtplib>=3.0.0",
    "email-validator>=2.1.0",
    "pillow>=10.0.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
]
